Make methods in times available at the top level of human_readable
This is a fix for #526. This makes methods in times available at the top level of human_readable.
Until now, `date_time` and `precise_delta` were imported, but the following methods were not available -- contrary to the content of the documentation.

- time_delta
- time_of_day
- timing
- year
- day
- date

This PR makes these methods available at the top level of human_readable.
It's just adding them to `__all__` in `human_readable/__init__.py`.

 # result

```python
>>> import human_readable
>>> human_readable.time_delta
<function time_delta at 0x1007bf420>
>>>
```